some reorganization and new files
