+++ src/game/character-init/initCharacter.js
