+++ src/game/character-init/initMoves.js
