+++ src/game/character-init/initStats.js
