+++ src/game/createCharInstance.js
@@ -1,4 +1,4 @@
-const initChar = require('./character-initialization/initCharacter')
+const initChar = require('./character-initn/initCharacter')
 
 module.exports = {
     initChar
