+++ src/game/overworld-init/initGamePiece.js
@@ -1,0 +1,15 @@
+function initGamePiece (
+    x, y, cellSize, type
+) {
+    this.x = x
+    this.y = y
+    this.width = cellSize,
+    this.height = cellSize,
+    this.sprite = getSprite(type),
+
+}
+
+function getSprite( type ) {
+    if // type
+    return // sprite
+}
